Add unit tests please

+++ src/main/java/interfaces/database/DBHandler.java
@@ -1,6 +1,5 @@
 package interfaces.database;
 
-import enums.UserRole;
 import models.Flight;
 import models.FlightFilter;
 
@@ -8,5 +7,5 @@ import java.util.List;
 
 public interface DBHandler {
     List<Flight> findFlights(FlightFilter filter);
-    Flight findFlightById(int id);
+    Flight findFlightById(Long id);
 }
